Přidána kontrola na odstraňované akcie - nesmí být nikde použité.

+++ app/model/StocksModel.php
@@ -9,6 +9,7 @@
 namespace App\Model;
 
 
+use App\Core\UIException;
 use App\DAL\Filters\StockFilter;
 use App\DAL\StocksRepo;
 
@@ -91,6 +92,7 @@ class StocksModel
 
     public function delete($id)
     {
+        if ($this->stocksRepo->isStockUsed($id)) throw new UIException("Tuto akcii nelze odstranit, protože je již použita (např. v uživatelském deníku).");
         $this->stocksRepo->deleteById($id);
     }
